pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "MCP를 이용하여 RAG mcp_tools 구현하기"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langgraph-cli[inmem]>=0.2.7",
    "mcp>=1.6.0",
    "pandas>=2.2.3",
    "setproctitle>=1.3.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
]

[tool.uv.workspace]
members = [
    "mcp_tools/rag",
    "mcp_tools/web_search",
    "mcp_tools/notion",
    "langchain_tools",
]
